pyproject.toml
[tool.poetry]
name = "svgdataset"
version = "0.1.0"
description = ""
authors = ["Umut Hope YILDIRIM <umut475@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyarrow = "^13.0.0"
pandas = "^1.3.4"
transformers = "^4.34.1"
pillow = "^10.1.0"
torch = "^2.1.0"
torchvision = "^0.16.0"
torchaudio = "^2.1.0"
cairosvg = "^2.7.1"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
